+++ packages/analytics/src/Analytics.ts
@@ -129,7 +129,6 @@ export class AnalyticsClass {
 			`The Analytics category has been configured successfully`
 		);
 		logger.debug('current configuration', this._config);
-
 		return this._config;
 	}
 
@@ -234,12 +233,6 @@ export class AnalyticsClass {
 		provider?,
 		metrics?: EventMetrics
 	) {
-		if (!this.isAnalyticsConfigured()) {
-			const errMsg = 'Analytics has not been configured';
-			logger.debug(errMsg);
-			return Promise.reject(new Error(errMsg));
-		}
-
 		let params = null;
 		// this is just for compatibility, going to be deprecated
 		if (typeof event === 'string') {
@@ -264,12 +257,6 @@ export class AnalyticsClass {
 	}
 
 	private _sendEvent(params) {
-		if (!this.isAnalyticsConfigured()) {
-			const errMsg = 'Analytics has not been configured';
-			logger.debug(errMsg);
-			return Promise.reject(new Error(errMsg));
-		}
-
 		if (this._disabled) {
 			logger.debug('Analytics has been disabled');
 			return Promise.resolve();
@@ -306,10 +293,6 @@ export class AnalyticsClass {
 		} else {
 			tracker.configure(opts);
 		}
-	}
-
-	private isAnalyticsConfigured() {
-		return this._config && Object.entries(this._config).length > 0;
 	}
 }