+++ packages/aws-amplify-react-native/src/Auth/SignUp.js
@@ -21,6 +21,7 @@ import {
 	Header,
 	ErrorRow,
 	AmplifyButton,
+	SignedOutMessage,
 	Wrapper,
 } from '../AmplifyUI';
 import AuthPiece from './AuthPiece';
@@ -59,7 +60,7 @@ export default class SignUp extends AuthPiece {
 	}
 
 	isValid() {
-		for (const el in this.signUpFields) {
+		for (const el of this.signUpFields) {
 			if (el.required && !this.state[el.key]) return false;
 		}
 		return true;
@@ -220,7 +221,7 @@ export default class SignUp extends AuthPiece {
 		this.sortFields();
 		return (
 			<Wrapper>
-				<ScrollView style={theme.section}>
+				<ScrollView style={theme.sectionScroll}>
 					<Header theme={theme} testID={TEST_ID.AUTH.SIGN_UP_TEXT}>
 						{I18n.get(this.header)}
 					</Header>
@@ -260,7 +261,7 @@ export default class SignUp extends AuthPiece {
 							text={I18n.get('Sign Up').toUpperCase()}
 							theme={theme}
 							onPress={this.signUp}
-							disabled={!this.isValid}
+							disabled={!this.isValid()}
 							testID={TEST_ID.AUTH.SIGN_UP_BUTTON}
 						/>
 					</View>
@@ -281,6 +282,7 @@ export default class SignUp extends AuthPiece {
 						</LinkCell>
 					</View>
 					<ErrorRow theme={theme}>{this.state.error}</ErrorRow>
+					<SignedOutMessage {...this.props} />
 				</ScrollView>
 			</Wrapper>
 		);